requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
passlib[bcrypt]
motor
PyJWT
email-validator
jinja2
